refactor(db): split schema by bounded context with backwards-compat barrel
The 567-line sip-and-speak.ts held 22 tables across all 5 bounded
contexts. Reading "the moderation tables" required scrolling. Splitting
the file makes the same locality wins as PR-1 did for routers.

Tables move into context-specific schema files:

  identity.ts     languageProfile, userLanguage, userInterest,
                  userDeviceToken (push tokens are user-bound infra)
  matching.ts     matchRequest, studentMatch
  scheduling.ts   venue, meetup, attendanceReport
  conversation.ts conversation, message, messageReadStatus,
                  messagingOptIn, conversationPresence
  moderation.ts   studentComment, blockedEmail, userFlag

relations() declarations live next to their table. Cross-context FKs
remain explicit (conversation.ts imports meetup from ./scheduling).

sip-and-speak.ts becomes a backwards-compat barrel:

  export * from "./identity";
  export * from "./matching";
  ...

Every existing import path resolves unchanged. No router, no notification
handler, no auth file changes — they all keep importing from
"@sip-and-speak/db/schema/sip-and-speak" and get the same exports.

No behavior change. Type errors: 6 unchanged (chat.ts pre-existing).
Tests: 206 pass / 31 fail across packages, identical to baseline.
Drizzle migration generation unaffected (same table definitions, same
output schema).
